#Update is an API method, not gits
