+++ git_test.go
@@ -1,9 +1,6 @@
 package main
 
-import (
-	"os"
-	"testing"
-)
+import "testing"
 
 func TestClonesValidRepo(t *testing.T) {
 	home := home()
@@ -47,34 +44,5 @@ func TestPullsRepo(t *testing.T) {
 	_, err := Pull(bundle, home)
 	if err != nil {
 		t.Error("No errors expected")
-	}
-}
-
-func TestUpdatesListOfRepos(t *testing.T) {
-	home := home()
-	bundle1 := "caarlos0/zsh-pg"
-	bundle2 := "caarlos0/zsh-add-upstream"
-	Clone(bundle1, home)
-	Clone(bundle2, home)
-	bundles, err := Update(home)
-	if err != nil {
-		t.Error("No errors expected")
-	}
-	if len(bundles) != 2 {
-		t.Error(len(bundles), "updated bundles, expected 2")
-	}
-}
-
-func TestUpdatesBrokenRepo(t *testing.T) {
-	home := home()
-	bundle := "caarlos0/zsh-pg"
-	folder, _ := Clone(bundle, home)
-	os.RemoveAll(folder + "/.git")
-	bundles, err := Update(home)
-	if err == nil {
-		t.Error("An error was expected")
-	}
-	if len(bundles) != 0 {
-		t.Error(len(bundles), "updated bundles, expected 0")
 	}
 }
